requirements.txt
asgiref==3.8.1
Django==5.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-static-fontawesome==6.5.2.0
pillow==10.4.0
psycopg2==2.9.9
python-environ==0.4.54
sqlparse==0.5.1
